refactor: lift the pure request, cookie and analytics helpers out of ipc-handlers
